src/main/java/com/toy/truffle/travel/controller/TravelController.java
package com.toy.truffle.travel.controller;

import com.toy.truffle.global.codeEnum.ResponseStatus;
import com.toy.truffle.global.dto.CommonResponseDTO;
import com.toy.truffle.travel.dto.TravelDto;
import com.toy.truffle.travel.service.TravelService;
import lombok.RequiredArgsConstructor;
import lombok.extern.slf4j.Slf4j;
import org.springframework.stereotype.Controller;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.ResponseBody;

@Controller
@Slf4j
@RequiredArgsConstructor
@RequestMapping("/travel")
public class TravelController {
    private final TravelService travelService;

    /**
     * 여행정보 저장
     */
    @PostMapping("/info")
    @ResponseBody
    public CommonResponseDTO saveTravel(@RequestBody TravelDto travelDto) {
        return travelService.saveTravel(travelDto);
    }
}
